+++ main/src/ca/uwaterloo/flix/language/phase/ClosureConv.scala
@@ -17,7 +17,7 @@
 package ca.uwaterloo.flix.language.phase
 
 import ca.uwaterloo.flix.language.GenSym
-import ca.uwaterloo.flix.language.ast.SimplifiedAst.{Expression, HandlerBinding}
+import ca.uwaterloo.flix.language.ast.SimplifiedAst.{CatchRule, Expression, HandlerBinding}
 import ca.uwaterloo.flix.language.ast.{Ast, SimplifiedAst, SourceLocation, Symbol, Type}
 import ca.uwaterloo.flix.util.InternalCompilerException
 import shapeless.ops.coproduct.ExtendRight
@@ -29,7 +29,7 @@ object ClosureConv {
   /**
     * Performs closure conversion on the given expression `e`.
     */
-  def convert(exp0: SimplifiedAst.Expression)(implicit genSym: GenSym): SimplifiedAst.Expression = exp0 match {
+  def visitExp(exp0: SimplifiedAst.Expression)(implicit genSym: GenSym): SimplifiedAst.Expression = exp0 match {
     case Expression.Unit => exp0
     case Expression.True => exp0
     case Expression.False => exp0
@@ -77,10 +77,10 @@ object ClosureConv {
 
       // Update the lambda type.
       val argTpes = freeVars.map(_._2) ++ targs
-      val newTpe = Type.mkArrow(argTpes, tresult)
+      val newTpe = Type.mkUncurriedArrow(argTpes, tresult)
 
       // We rewrite the lambda with its new arguments list and new body, with any nested lambdas also converted.
-      val lambda = Expression.Lambda(newArgs, convert(replace(body, subst.toMap)), newTpe, loc)
+      val lambda = Expression.Lambda(newArgs, visitExp(replace(body, subst.toMap)), newTpe, loc)
 
       // At this point, `lambda` is the original lambda expression, but with all free variables converted to new
       // arguments, prepended to the original arguments list. Additionally, any lambdas within the body have also been
@@ -99,25 +99,25 @@ object ClosureConv {
       // it with ApplyRef. We remove the Ref node and don't recurse on it to avoid creating a closure.
       // We do something similar if `e` is a Hook, where we transform Apply to ApplyHook.
       e match {
-        case Expression.Def(sym, _, _) => Expression.ApplyDef(sym, args.map(convert), tpe, loc)
-        case Expression.Eff(sym, _, _) => Expression.ApplyEff(sym, args.map(convert), tpe, loc)
-        case _ => Expression.ApplyClo(convert(e), args.map(convert), tpe, loc)
+        case Expression.Def(sym, _, _) => Expression.ApplyDef(sym, args.map(visitExp), tpe, loc)
+        case Expression.Eff(sym, _, _) => Expression.ApplyEff(sym, args.map(visitExp), tpe, loc)
+        case _ => Expression.ApplyClo(visitExp(e), args.map(visitExp), tpe, loc)
       }
 
 
     case Expression.Unary(sop, op, e, tpe, loc) =>
-      Expression.Unary(sop, op, convert(e), tpe, loc)
+      Expression.Unary(sop, op, visitExp(e), tpe, loc)
 
     case Expression.Binary(sop, op, e1, e2, tpe, loc) =>
-      Expression.Binary(sop, op, convert(e1), convert(e2), tpe, loc)
+      Expression.Binary(sop, op, visitExp(e1), visitExp(e2), tpe, loc)
 
     case Expression.IfThenElse(e1, e2, e3, tpe, loc) =>
-      Expression.IfThenElse(convert(e1), convert(e2), convert(e3), tpe, loc)
+      Expression.IfThenElse(visitExp(e1), visitExp(e2), visitExp(e3), tpe, loc)
 
     case Expression.Branch(exp, branches, tpe, loc) =>
-      val e = convert(exp)
+      val e = visitExp(exp)
       val bs = branches map {
-        case (sym, br) => sym -> convert(br)
+        case (sym, br) => sym -> visitExp(br)
       }
       Expression.Branch(e, bs, tpe, loc)
 
@@ -125,83 +125,99 @@ object ClosureConv {
       Expression.JumpTo(sym, tpe, loc)
 
     case Expression.Let(sym, e1, e2, tpe, loc) =>
-      Expression.Let(sym, convert(e1), convert(e2), tpe, loc)
+      Expression.Let(sym, visitExp(e1), visitExp(e2), tpe, loc)
 
     case Expression.LetRec(sym, e1, e2, tpe, loc) =>
-      Expression.LetRec(sym, convert(e1), convert(e2), tpe, loc)
+      Expression.LetRec(sym, visitExp(e1), visitExp(e2), tpe, loc)
 
     case Expression.Is(sym, tag, e, loc) =>
-      Expression.Is(sym, tag, convert(e), loc)
+      Expression.Is(sym, tag, visitExp(e), loc)
 
     case Expression.Tag(enum, tag, e, tpe, loc) =>
-      Expression.Tag(enum, tag, convert(e), tpe, loc)
+      Expression.Tag(enum, tag, visitExp(e), tpe, loc)
 
     case Expression.Untag(sym, tag, e, tpe, loc) =>
-      Expression.Untag(sym, tag, convert(e), tpe, loc)
+      Expression.Untag(sym, tag, visitExp(e), tpe, loc)
 
     case Expression.Index(e, offset, tpe, loc) =>
-      Expression.Index(convert(e), offset, tpe, loc)
+      Expression.Index(visitExp(e), offset, tpe, loc)
 
     case Expression.Tuple(elms, tpe, loc) =>
-      Expression.Tuple(elms.map(convert), tpe, loc)
+      Expression.Tuple(elms.map(visitExp), tpe, loc)
 
     case Expression.ArrayLit(elms, tpe, loc) =>
-      Expression.ArrayLit(elms.map(convert), tpe, loc)
+      Expression.ArrayLit(elms.map(visitExp), tpe, loc)
 
     case Expression.ArrayNew(elm, len, tpe, loc) =>
-      val e1 = convert(elm)
-      val e2 = convert(len)
+      val e1 = visitExp(elm)
+      val e2 = visitExp(len)
       Expression.ArrayNew(e1, e2, tpe, loc)
 
-    case Expression.ArrayLoad(exp1, exp2, tpe, loc) =>
-      val e1 = convert(exp1)
-      val e2 = convert(exp2)
+    case Expression.ArrayLoad(base, index, tpe, loc) =>
+      val e1 = visitExp(base)
+      val e2 = visitExp(index)
       Expression.ArrayLoad(e1, e2, tpe, loc)
 
-    case Expression.ArrayStore(exp1, exp2, exp3, tpe, loc) =>
-      val e1 = convert(exp1)
-      val e2 = convert(exp2)
-      val e3 = convert(exp3)
+    case Expression.ArrayStore(base, index, elm, tpe, loc) =>
+      val e1 = visitExp(base)
+      val e2 = visitExp(index)
+      val e3 = visitExp(elm)
       Expression.ArrayStore(e1, e2, e3, tpe, loc)
 
-    case Expression.ArrayLength(exp, tpe, loc) =>
-      val e = convert(exp)
+    case Expression.ArrayLength(base, tpe, loc) =>
+      val e = visitExp(base)
       Expression.ArrayLength(e, tpe, loc)
 
-    case Expression.ArraySlice(exp1, exp2, exp3, tpe, loc) =>
-      val e1 = convert(exp1)
-      val e2 = convert(exp2)
-      val e3 = convert(exp3)
+    case Expression.ArraySlice(base, beginIndex, endIndex, tpe, loc) =>
+      val e1 = visitExp(base)
+      val e2 = visitExp(beginIndex)
+      val e3 = visitExp(endIndex)
       Expression.ArraySlice(e1, e2, e3, tpe, loc)
 
     case Expression.Ref(exp, tpe, loc) =>
-      val e = convert(exp)
+      val e = visitExp(exp)
       Expression.Ref(e, tpe, loc)
 
     case Expression.Deref(exp, tpe, loc) =>
-      val e = convert(exp)
+      val e = visitExp(exp)
       Expression.Deref(e, tpe, loc)
 
     case Expression.Assign(exp1, exp2, tpe, loc) =>
-      val e1 = convert(exp1)
-      val e2 = convert(exp2)
+      val e1 = visitExp(exp1)
+      val e2 = visitExp(exp2)
       Expression.Assign(e1, e2, tpe, loc)
 
     case Expression.HandleWith(exp, bindings, tpe, loc) =>
-      val e = convert(exp)
+      val e = visitExp(exp)
       val bs = bindings map {
-        case HandlerBinding(sym, handler) => HandlerBinding(sym, convert(handler))
+        case HandlerBinding(sym, handler) => HandlerBinding(sym, visitExp(handler))
       }
       Expression.HandleWith(e, bs, tpe, loc)
 
     case Expression.Existential(params, e, loc) =>
-      Expression.Existential(params, convert(e), loc)
-    case Expression.Universal(params, e, loc) =>
-      Expression.Universal(params, convert(e), loc)
+      Expression.Existential(params, visitExp(e), loc)
 
-    case Expression.NativeConstructor(constructor, args, tpe, loc) => exp0
+    case Expression.Universal(params, e, loc) =>
+      Expression.Universal(params, visitExp(e), loc)
+
+    case Expression.TryCatch(exp, rules, tpe, eff, loc) =>
+      val e = visitExp(exp)
+      val rs = rules map {
+        case CatchRule(sym, clazz, body) =>
+          val b = visitExp(body)
+          CatchRule(sym, clazz, b)
+      }
+      Expression.TryCatch(e, rs, tpe, eff, loc)
+
+    case Expression.NativeConstructor(constructor, args, tpe, loc) =>
+      val as = args map visitExp
+      Expression.NativeConstructor(constructor, as, tpe, loc)
+
     case Expression.NativeField(field, tpe, loc) => exp0
-    case Expression.NativeMethod(method, args, tpe, loc) => exp0
+
+    case Expression.NativeMethod(method, args, tpe, loc) =>
+      val as = args map visitExp
+      Expression.NativeMethod(method, as, tpe, loc)
 
     case Expression.UserError(tpe, loc) => exp0
     case Expression.HoleError(sym, tpe, eff, loc) => exp0
@@ -280,6 +296,8 @@ object ClosureConv {
       freeVariables(exp).filterNot { v => v._1 == fparam.sym }
     case Expression.Universal(fparam, exp, loc) =>
       freeVariables(exp).filterNot { v => v._1 == fparam.sym }
+
+    case Expression.TryCatch(exp, rules, tpe, eff, loc) => mutable.LinkedHashSet.empty ++ freeVariables(exp) ++ rules.flatMap(r => freeVariables(r.exp).filterNot(_._1 == r.sym))
     case Expression.NativeConstructor(constructor, args, tpe, loc) => mutable.LinkedHashSet.empty ++ args.flatMap(freeVariables)
     case Expression.NativeField(field, tpe, loc) => mutable.LinkedHashSet.empty
     case Expression.NativeMethod(method, args, tpe, loc) => mutable.LinkedHashSet.empty ++ args.flatMap(freeVariables)
@@ -442,10 +460,22 @@ object ClosureConv {
         val fs = replace(fparam, subst)
         val e = visit(exp)
         Expression.Universal(fs, e, loc)
+
+      case Expression.TryCatch(exp, rules, tpe, eff, loc) =>
+        val e = visit(exp)
+        val rs = rules map {
+          case CatchRule(sym, clazz, body) =>
+            val b = visit(body)
+            CatchRule(sym, clazz, b)
+        }
+        Expression.TryCatch(e, rs, tpe, eff, loc)
+
       case Expression.NativeConstructor(constructor, args, tpe, loc) =>
         val es = args map visit
         Expression.NativeConstructor(constructor, es, tpe, loc)
+
       case Expression.NativeField(field, tpe, loc) => e
+
       case Expression.NativeMethod(method, args, tpe, loc) =>
         val es = args map visit
         Expression.NativeMethod(method, es, tpe, loc)